ui: fix import cycles in various modules
