+++ ui/analyse/src/api.ts
@@ -1,5 +1,5 @@
 import * as control from './control';
-import AnalyseCtrl from './ctrl';
+import type AnalyseCtrl from './ctrl';
 
 export default function (ctrl: AnalyseCtrl) {
   return {
