+++ ui/analyse/src/retrospect/nvuiRetroView.ts
@@ -4,10 +4,10 @@ import { renderSan } from 'lib/nvui/chess';
 import { liveText } from 'lib/nvui/notify';
 import { type LooseVNodes, hl } from 'lib/view';
 
-import type { AnalyseNvuiContext } from '../analyse.nvui';
-import type AnalyseCtrl from '../ctrl';
-import type { RetroCtrl } from '../retrospect/retroCtrl';
-import { clickHook, renderCurrentNode } from '../view/nvuiView';
+import type { AnalyseNvuiContext } from '@/analyse.nvui';
+import type AnalyseCtrl from '@/ctrl';
+import { clickHook, renderCurrentNode } from '@/nvuiUtil';
+import type { RetroCtrl } from '@/retrospect/retroCtrl';
 
 export function renderRetro(nvuiCtx: AnalyseNvuiContext): LooseVNodes {
   const ctx = makeContext(nvuiCtx);
