+++ ui/analyse/src/study/relay/relayTeamLeaderboard.ts
@@ -2,7 +2,7 @@ import type { Tablesort } from 'tablesort';
 
 import { throttle } from 'lib';
 import { Group, StudyBoard } from 'lib/licon';
-import { dataIcon, hl, onInsert, requiresI18n, spinnerVdom, type VNode, type VNodeData } from 'lib/view';
+import { dataIcon, hl, onInsert, requiresI18n, spinnerVdom, type VNode } from 'lib/view';
 import { json as xhrJson } from 'lib/xhr';
 
 import type { StudyPlayerFromServer } from '../interfaces';
@@ -187,9 +187,3 @@ export default class RelayTeamLeaderboard {
     this.setTeamToShow(team);
   };
 }
-
-export const teamLinkData = (teamName: RelayTeamName): VNodeData => ({
-  attrs: {
-    href: `#team-results/${encodeURIComponent(teamName)}`,
-  },
-});
