+++ ui/analyse/src/study/relay/relayTeams.ts
@@ -2,14 +2,15 @@ import { type MaybeVNodes, type VNode, onInsert, hl, spinnerVdom as spinner } fr
 import { userTitle } from 'lib/view/userLink';
 import { json as xhrJson } from 'lib/xhr';
 
+import { playerFedFlag } from '@/view/util';
+
 import type { ChapterId, ChapterPreview, StudyPlayer, ChapterSelect } from '../interfaces';
 import { type MultiCloudEval, renderScore } from '../multiCloudEval';
-import { playerFedFlag } from '../playerBars';
 import { gameLinkAttrs, gameLinksListener, StudyChapters } from '../studyChapters';
 import { coloredStatusStr } from './customScoreStatus';
+import { teamLinkData } from './deepLink';
 import type { RelayRound, RelayTour } from './interfaces';
 import type RelayPlayers from './relayPlayers';
-import { teamLinkData } from './relayTeamLeaderboard';
 
 interface TeamWithPoints {
   name: string;
