+++ ui/analyse/src/study/relay/relayView.ts
@@ -1,8 +1,7 @@
-import { onClickAway } from 'lib';
 import { view as cevalView } from 'lib/ceval';
 import { displayColumns, isTouchDevice } from 'lib/device';
 import * as licon from 'lib/licon';
-import { bind, dataIcon, hl, onInsert, type VNode } from 'lib/view';
+import { bind, dataIcon, hl, type VNode } from 'lib/view';
 
 import type AnalyseCtrl from '@/ctrl';
 import { view as keyboardView } from '@/keyboard';
@@ -11,14 +10,15 @@ import {
   viewContext,
   renderBoard,
   renderMain,
-  renderTools,
   renderUnderboard,
 } from '@/view/components';
 import { renderControls } from '@/view/controls';
+import { renderTools } from '@/view/tools';
 
 import type * as studyDeps from '../studyDeps';
 import type RelayCtrl from './relayCtrl';
 import { tourSide, renderRelayTour } from './relayTourView';
+import { allowVideo } from './videoPlayer';
 
 export function relayView(
   ctrl: AnalyseCtrl,
@@ -64,37 +64,6 @@ export const backToLiveView = (ctrl: AnalyseCtrl) =>
         i18n.broadcast.backToLiveMove,
       )
     : undefined;
-
-export function renderStreamerMenu(relay: RelayCtrl): VNode {
-  const makeUrl = (id: string) => {
-    const url = new URL(location.href);
-    url.searchParams.set('embed', id);
-    return url.toString();
-  };
-  return hl(
-    'div.streamer-menu-anchor',
-    hl(
-      'div.streamer-menu',
-      {
-        hook: onInsert(
-          onClickAway(() => {
-            relay.showStreamerMenu(false);
-            relay.redraw();
-          }),
-        ),
-      },
-      relay.streams.map(([id, info]) =>
-        hl('a.streamer.text', { attrs: { 'data-icon': licon.Mic, href: makeUrl(id) } }, [
-          info.name,
-          hl('i', info.lang),
-        ]),
-      ),
-    ),
-  );
-}
-
-export const allowVideo = (): boolean =>
-  window.getComputedStyle(document.body).getPropertyValue('---allow-video') === 'true';
 
 function renderBoardView(ctx: RelayViewContext) {
   const { ctrl, deps, study, gaugeOn, relay } = ctx;
