+++ ui/analyse/src/study/relay/videoPlayer.ts
@@ -1,7 +1,5 @@
 import { hl, type VNode, onInsert } from 'lib/view';
 
-import { allowVideo } from './relayView';
-
 export class VideoPlayer {
   private iframe: HTMLIFrameElement;
   private close: HTMLImageElement;
@@ -122,3 +120,6 @@ export class VideoPlayer {
     window.location.href = urlWithEmbed.toString();
   };
 }
+
+export const allowVideo = (): boolean =>
+  window.getComputedStyle(document.body).getPropertyValue('---allow-video') === 'true';
