+++ ui/analyse/src/view/materialDiffs.ts
@@ -1,0 +1,15 @@
+import type { VNode } from 'snabbdom';
+
+import * as materialView from 'lib/game/view/material';
+
+import type AnalyseCtrl from '@/ctrl';
+
+export const renderMaterialDiffs = (ctrl: AnalyseCtrl): [VNode, VNode] =>
+  materialView.renderMaterialDiffs(
+    !!ctrl.data.pref.showCaptured,
+    ctrl.bottomColor(),
+    ctrl.node.fen,
+    !!(ctrl.data.player.checks || ctrl.data.opponent.checks), // showChecks
+    ctrl.nodeList,
+    ctrl.node.ply,
+  );
