+++ ui/analyse/src/view/nvuiView.ts
@@ -39,7 +39,7 @@ import { renderSetting } from 'lib/nvui/setting';
 import { pubsub } from 'lib/pubsub';
 import { ops, path as treePath } from 'lib/tree/tree';
 import type { ClientEval, PvData } from 'lib/tree/types';
-import { type VNode, type LooseVNodes, type VNodeChildren, hl, bind, noTrans, enter } from 'lib/view';
+import { type VNode, type LooseVNodes, type VNodeChildren, hl, bind, noTrans } from 'lib/view';
 import { text as xhrText } from 'lib/xhr';
 
 import type { AnalyseNvuiContext } from '../analyse.nvui';
@@ -48,6 +48,7 @@ import type AnalyseCtrl from '../ctrl';
 import explorerView from '../explorer/explorerView';
 import { makeConfig as makeCgConfig } from '../ground';
 import type { AnalyseData } from '../interfaces';
+import { clickHook, currentLineIndex, renderCurrentNode } from '../nvuiUtil';
 import { renderRetro } from '../retrospect/nvuiRetroView';
 import { view as chapterEditFormView } from '../study/chapterEditForm';
 import { view as chapterNewFormView } from '../study/chapterNewForm';
@@ -216,29 +217,6 @@ export function renderNvui(ctx: AnalyseNvuiContext): VNode {
       deps && ctrl.study?.relay && tourDetails(ctx),
     ]),
   ]);
-}
-
-export function clickHook(main?: (el: HTMLElement) => void, post?: () => void) {
-  return {
-    // put unique identifying props on the button container (such as class)
-    // because snabbdom WILL mix plain adjacent buttons up.
-    hook: {
-      insert: (vnode: VNode) => {
-        const el = vnode.elm as HTMLElement;
-        el.addEventListener('click', () => {
-          main?.(el);
-          post?.();
-        });
-        el.addEventListener(
-          'keydown',
-          enter(() => {
-            main?.(el);
-            post?.();
-          }),
-        );
-      },
-    },
-  };
 }
 
 function boardEventsHook(
@@ -513,38 +491,6 @@ const requestAnalysisBtn = ({ ctrl, notify, analysisInProgress }: AnalyseNvuiCon
         i18n.site.requestAComputerAnalysis,
       );
 };
-
-function currentLineIndex(ctrl: AnalyseCtrl): { i: number; of: number } {
-  if (ctrl.path === treePath.root) return { i: 1, of: 1 };
-  const prevNode = ctrl.tree.parentNode(ctrl.path);
-  return {
-    i: prevNode.children.findIndex(node => node.id === ctrl.node.id),
-    of: prevNode.children.length,
-  };
-}
-
-function renderLineIndex(ctrl: AnalyseCtrl): string {
-  const { i, of } = currentLineIndex(ctrl);
-  return of > 1 ? `, line ${i + 1} of ${of} ,` : '';
-}
-
-export function renderCurrentNode({
-  ctrl,
-  moveStyle,
-}: Pick<AnalyseNvuiContext, 'ctrl' | 'moveStyle'>): string {
-  const node = ctrl.node;
-  if (!node.san || !node.uci) return i18n.nvui.gameStart;
-  return [
-    plyToTurn(node.ply),
-    node.ply % 2 === 1 ? i18n.site.white : i18n.site.black,
-    renderSan(node.san, node.uci, moveStyle.get()),
-    renderLineIndex(ctrl),
-    !ctrl.retro && renderComments(node, moveStyle.get()),
-  ]
-    .filter(x => x)
-    .join(' ')
-    .trim();
-}
 
 const renderPlayer = (ctrl: AnalyseCtrl, player: Player): LooseVNodes =>
   player.ai ? i18n.site.aiNameLevelAiLevel('Stockfish', player.ai) : userHtml(ctrl, player);
