+++ ui/botPlay/src/play/sound.ts
@@ -2,7 +2,7 @@ import { type SoundEvent } from 'lib/bot/types';
 
 import type { Move } from '@/interfaces';
 
-import PlayCtrl from './playCtrl';
+import type PlayCtrl from './playCtrl';
 
 export const playMoveSounds = async (ctrl: PlayCtrl, move: Move) => {
   const san = move.san;
