+++ ui/dasher/src/interfaces.ts
@@ -1,10 +1,8 @@
 import type { VNode } from 'lib/view';
 
-import type { BackgroundData } from './background';
-import { DasherCtrl } from './ctrl';
-import type { LangsData } from './langs';
-
-export { DasherCtrl };
+import type { BackgroundData } from '@/background';
+import type { DasherCtrl } from '@/ctrl';
+import type { LangsData } from '@/langs';
 
 export type Dimension = 'd2' | 'd3';
