+++ ui/dasher/src/ping.ts
@@ -3,7 +3,7 @@ import { h, type VNode } from 'snabbdom';
 import { defined } from 'lib';
 import { pubsub } from 'lib/pubsub';
 
-import type { DasherCtrl } from './interfaces';
+import type { DasherCtrl } from '@/ctrl';
 
 export class PingCtrl {
   ping: number | undefined;
