+++ ui/insight/src/insight.refresh.ts
@@ -1,3 +1,14 @@
-import { registerFormHandler } from './insight';
+export function registerFormHandler() {
+  $('form.insight-refresh').on('submit', function (this: HTMLFormElement) {
+    fetch(this.action, {
+      method: 'post',
+      credentials: 'same-origin',
+    }).then(site.reload);
+
+    $(this).replaceWith($(this).find('.crunching').removeClass('none'));
+
+    return false;
+  });
+}
 
 site.load.then(registerFormHandler);
