+++ ui/insight/src/insight.ts
@@ -26,16 +26,3 @@ export function initModule(opts: Env) {
 
   return ctrl;
 }
-
-export function registerFormHandler() {
-  $('form.insight-refresh').on('submit', function (this: HTMLFormElement) {
-    fetch(this.action, {
-      method: 'post',
-      credentials: 'same-origin',
-    }).then(site.reload);
-
-    $(this).replaceWith($(this).find('.crunching').removeClass('none'));
-
-    return false;
-  });
-}
