+++ ui/voice/src/util.ts
@@ -123,3 +123,9 @@ export function dest(uci: Uci) {
 }
 
 export const promo = (uci: Uci): Role | undefined => charToRole(uci.slice(4, 5));
+
+export function flash(): void {
+  const div = document.querySelector<HTMLElement>('#voice-status-row')!;
+  div.classList.add('flash');
+  div.onanimationend = () => div.classList.remove('flash');
+}
