+++ ui/voice/src/voice.ts
@@ -10,27 +10,11 @@ import {
 import type { VoiceCtrl, VoiceModule } from './interfaces';
 import { Mic } from './mic';
 import type { VoiceMove } from './move/interfaces';
-import { flash } from './view';
+import { flash } from './util';
 
 export * from './interfaces';
 export * from './move/interfaces';
 export { renderVoiceBar } from './view';
-
-export const supportedLangs: [string, string][] = [
-  ['en', 'English'],
-  ['fr', 'Français'],
-  ['pl', 'Polski'],
-];
-
-if (site.debug)
-  supportedLangs.push(
-    ['de', 'Deutsch'],
-    ['tr', 'Türkçe'],
-    ['vi', 'Tiếng Việt'],
-    ['ru', 'Русский'],
-    ['it', 'Italiano'],
-    ['sv', 'Svenska'],
-  );
 
 export function makeVoice(opts: {
   redraw: () => void;
